feat: overhaul Dark Matter economy (Full Stack)
Boosted passive generation. Implemented Matter Synthesis with 30% fee (10k:0.7 ratio) on Desktop and Mobile. Added auto-comma input masking. Updated Void Container loot. Updated DB schema to DECIMAL(40,18) for safe high-precision storage.

+++ app/Models/Repositories/ResourceRepository.php
@@ -128,10 +128,10 @@ soldiers: (int)$data['soldiers'],
 guards: (int)$data['guards'],
 spies: (int)$data['spies'],
 sentries: (int)$data['sentries'],
-untraceable_chips: (int)($data['untraceable_chips'] ?? 0),
-research_data: (int)($data['research_data'] ?? 0),
-dark_matter: (int)($data['dark_matter'] ?? 0),
-protoform: (float)($data['protoform'] ?? 0.0)
-);
-}
+            untraceable_chips: (int)($data['untraceable_chips'] ?? 0),
+            research_data: (int)($data['research_data'] ?? 0),
+            dark_matter: (float)($data['dark_matter'] ?? 0.0),
+            protoform: (float)($data['protoform'] ?? 0.0)
+        );
+    }
 }
